move export functions from normalizers.utils to utils
